profile changes

+++ 3d-design-app/src/datalayer/querys.ts
@@ -1,6 +1,17 @@
 import { db } from "./config"
 import { addDoc, collection, getDocs, query, where, doc, setDoc, deleteDoc, updateDoc } from "firebase/firestore"
 
+function generateId(length: number) {
+    let id = ''
+    let chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
+
+    for (let i = 0; i < length; i++) {
+        id += chars.charAt(Math.floor(Math.random() * chars.length))
+    }
+
+    return id
+}
+
 async function getUserData(data: any) {
     const que = query(collection(db, 'data'), where('userId', '==', data.userId))
     const querySnapshot = await getDocs(que)
@@ -10,6 +21,8 @@ async function getUserData(data: any) {
 }
 
 async function checkUser(userId: string, userName: string, userEmail: string, method: string) {
+    const messagingId = generateId(5)
+
     const que = query(collection(db, 'data'), where('userId', '==', userId))
     const querySnapshot = await getDocs(que)
 
@@ -19,7 +32,9 @@ async function checkUser(userId: string, userName: string, userEmail: string, me
                 userId: userId,
                 username: userName,
                 email: userEmail,
-                method: method
+                method: method,
+                locked: false,
+                messagingId: messagingId
             })
 
             console.log(`Data added with id: ${docRef.id}`);
@@ -47,12 +62,7 @@ async function cookieSetter(userState: boolean, userId: string | null) {
 
 
 async function updateDesign(userId: string, action: string, id: string, newName: string) {
-    let desId = ''
-    let chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
-
-    for (let i = 0; i < 8; i++) {
-        desId += chars.charAt(Math.floor(Math.random() * chars.length))
-    }
+    const desId = generateId(8)
 
     const que = query(collection(db, 'data'), where('userId', '==', userId))
     const querySnapshot = await getDocs(que)
@@ -83,21 +93,41 @@ async function updateDesign(userId: string, action: string, id: string, newName:
     }
 }
 
-async function updateFriend(userId: string, action: string, friendId: string, friendName: string) {
+async function updateFriendOrUser(userId: string, action: string, friendId: string | any, friendName: string | any, messagingId: string | any, friendOrUser: string, state: boolean | null) {
     const que = query(collection(db, 'data'), where('userId', '==', userId))
     const querySnapshot = await getDocs(que)
     const docId = querySnapshot.docs.map((doc) => doc.id)
 
-    if (action === 'remove') {
+    if (action === 'remove' && friendOrUser === 'friend') {
         const docRef = doc(db, 'data', docId[0], 'friends', friendId)
         await deleteDoc(docRef)
+
+    } else if (action === 'update' && friendOrUser === 'user') {
+        const docRef = doc(db, 'data', docId[0])
+        await updateDoc(docRef, {
+            'locked': state
+        })
+
+    } else if (action === 'add' && friendOrUser === 'friend') {
+        const docRef = doc(db, 'data', docId[0], 'friends', friendId)
+        await setDoc(docRef, {
+            friendData: {
+                friendId: friendId,
+                friendName: friendName,
+                messagingId: messagingId
+            },
+            user: userId
+        })
     }
 }
+
+
 
 export {
     getUserData,
     checkUser,
     cookieSetter,
     updateDesign,
-    updateFriend
+    updateFriendOrUser,
+    generateId
 }
